Formatted agEmbeddedAgAddressValueForm.
Added a check in agPerson's _call method that determines whether or not the desired results from the helper method have been returned. If they have, it returns them, if they haven't, it returns nothing. This stopped an undefined index warning on the staff show page that resulted from a null set being returned by the address helper.

Added a check in staff actions to determine if the name types being used to create the page's title actually exist. If they do, the title gets them. If not, the title gets nothing. This prevented another undefined index error on the staff show page.

Added code to staff showSuccess to display feedback indicating that the staff person has no address information inside of the address table, rather than just displaying the address table header without and table content.