The modal-modal is being changed to a slide down in the original modal. Much nicer looking, less confusing to the user, and other greate stuff.
Some refactoring of code in various event templates to increase readability.

Added a check to the listgroups event action that will determine whether an event passed in the url is an actual event. If not, it redirects to listgroups without an event.

Commented out the check describe in the sentence above, since having no event in the url on listgroups will break the functionality of groupdetail. It's there for when we fix it later.

Added a check to listgroups to see whether or not an event is in it's url. If it isn't you get a warning and directions to go somewhere else.