A lot of refactor of the agPersonForm.class.php. Lat and long are now fully CRUDable through the staff form front end. The facility form has fields for lat and long, but does not save yet. After the code is tested more on staff, it will be proted to facility.
The only change to agGeoHelper was to make the setDefaultGeoMatchScoreId() function set the _defaultGeoMatchScoreId property before completing.

A strtolower was removed from the agBulkRecordHelper ucTrim() function, but the code has been reverted since address saving is no longer relying on this class (the strtolower was causing all address values to become lowercased). The strtolower can easily be removed though, if it need to be removed for any other reason.

The rest is just some tweaking of forms and formatters and the addition of some classes to the css.